Extracted editable lists and task into separate comps
